Blinded agents to neighbors and positions.
Agent proposal functions can no longer see the full list of neighbors
(only those sensed) or the agent's position (which will appear as [0 0]
if accessed in the proposal function). The position of an agent in a
spawn proposal will be offset relative to the parent's position when
the child agent is spawned.