Changed :draw-functions to :draw-function.
Previously each agent could have a sequence of draw functions, all of
which would be called. Since this wasn't being used, I made the
simplification that each agent would have just one draw-function. So
now instead of having :draw-functions an agent has a :draw-function,
the value of which should be a function that takes the agent as an
argument.